requirements.txt
Babel==2.5.3
Django==2.0.4
django-crispy-forms==1.7.2
django-extra-views==0.11.0
django-phonenumber-field==2.0.0
phonenumberslite==8.9.4
Pillow==5.1.0
pkg-resources==0.0.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
pytz==2018.4
six==1.11.0
